# Additional cleanup and acceleration of the System.Numerics vector and related types