--- a/src/libraries/System.Private.CoreLib/src/System/Numerics/Vector3.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Numerics/Vector3.cs
@@ -302,6 +302,16 @@ namespace System.Numerics
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Vector3 Cross(Vector3 vector1, Vector3 vector2)
         {
+            if (Vector128.IsHardwareAccelerated)
+            {
+                var v1 = vector1.AsVector128();
+                var v2 = vector2.AsVector128();
+                return (Vector128.Shuffle(v1, Vector128.Create(1, 2, 0, 3)) *
+                        Vector128.Shuffle(v2, Vector128.Create(2, 0, 1, 3))) -
+                       (Vector128.Shuffle(v1, Vector128.Create(2, 0, 1, 3)) *
+                        Vector128.Shuffle(v2, Vector128.Create(1, 2, 0, 3)));
+            }
+            
             return new Vector3(
                 (vector1.Y * vector2.Z) - (vector1.Z * vector2.Y),
                 (vector1.Z * vector2.X) - (vector1.X * vector2.Z),
@@ -516,6 +526,23 @@ namespace System.Numerics
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Vector3 Transform(Vector3 value, Quaternion rotation)
         {
+            if (Vector128.IsHardwareAccelerated)
+            {
+                var vVector = value.AsVector128();
+                var rVector = Unsafe.BitCast<Quaternion, Vector128<float>>(rotation);
+
+                // v + 2 * (q x (q.W  * v + q x v)
+                return (vVector + Vector128.Create(2f) * Cross(qVector, Vector128.Shuffle(qVector, Vector128.Create(3, 3, 3, 3)) * vVector + Cross(qVector, vVector))).AsVector3();
+
+                static Vector128<float> Cross(Vector128<float> v1, Vector128<float> v2)
+                {
+                    return (Vector128.Shuffle(v1, Vector128.Create(1, 2, 0, 3)) *
+                            Vector128.Shuffle(v2, Vector128.Create(2, 0, 1, 3))) -
+                           (Vector128.Shuffle(v1, Vector128.Create(2, 0, 1, 3)) *
+                            Vector128.Shuffle(v2, Vector128.Create(1, 2, 0, 3)));
+                }
+            }
+
             float x2 = rotation.X + rotation.X;
             float y2 = rotation.Y + rotation.Y;
             float z2 = rotation.Z + rotation.Z;