… src/libraries/System.Numerics.Vectors/tests/QuaternionTests.cs
… src/libraries/System.Numerics.Vectors/tests/Vector2Tests.cs
… src/libraries/System.Numerics.Vectors/tests/Vector3Tests.cs
… src/libraries/System.Numerics.Vectors/tests/Vector4Tests.cs
… src/libraries/System.Private.CoreLib/src/System/Numerics/Vector2.cs
… src/libraries/System.Private.CoreLib/src/System/Numerics/Vector4.cs